+++ flutter_quill_extensions/lib/embeds/video/toolbar/video_button.dart
@@ -2,6 +2,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_quill/flutter_quill.dart';
+import 'package:flutter_quill/translations.dart';
 
 import '../../../models/config/shared_configurations.dart';
 import '../../../models/config/toolbar/buttons/video.dart';