+++ lib/translations.dart
@@ -1,3 +1,4 @@
 library flutter_quill.translations;
 
 export 'src/l10n/extensions/localizations.dart';
+export 'src/l10n/widgets/localizations.dart';